refactor: CloudStorageService 인터페이스+구현체로 변경
유연성과 확장성을 위해서 변경했습니다.
